fix: keep infrastructure out of input factory
