+++ packages/riviere-extract-ts/use-cases/src/features/extract/commands/extract-draft-components-input.ts
@@ -1,4 +1,8 @@
 /** @riviere-role command-use-case-input */
+export type SourceFileSelection =
+  | { readonly kind: 'all' }
+  | { readonly kind: 'files'; readonly filePaths: readonly string[] }
+
 export interface ExtractDraftComponentsInput {
   allowIncomplete: boolean
   baseBranch?: string
@@ -6,9 +10,7 @@ export interface ExtractDraftComponentsInput {
   files?: string[]
   includeConnections: boolean
   projectRoot?: string
-  sourceFileSelection:
-    | { readonly kind: 'all' }
-    | { readonly kind: 'files'; readonly filePaths: readonly string[] }
+  sourceFileSelection: SourceFileSelection
   output?: string
   sourceMode: 'all' | 'files' | 'pull-request'
   useTsConfig: boolean
